requirements.txt
Django==3.2
psycopg2
requests
selenium
bs4
Pillow